Cake Support
Add support for Cake (version 0.12.05) as a stand-alone preprocessor that can convert C99, C11 & C23 programs to C89, suitable for compilation by Catalina. Cake is supported on Windows, Linux and Catalyst.

Fix an issue with sprintf and vsprintf which limed output of any single call to 32k

Fix some issues with the Catalyst build_demos and build_p2 scripts for Linux.